hide.h.m4: hide_functor::operator(): Make this variadic.
This uses some tuple manipulation utilities so that the variadic parameters
with which we call a method can be based on the variadic parameters that the
caller received, but not exactly. In this case, we need to replace an element
in the middle of a tuple. There is probably a more efficient way to do this.

For now, this is using copies of the tuple utilities from here:
https://github.com/murraycu/murrayc-tuple-utils/tree/master/tuple-utils